day-19-resume-screener-agent/pyproject.toml
[project]
name = "day-19-resume-screener-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.14",
    "langgraph>=1.0.1",
    "openai>=1.109.1",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "uvicorn>=0.41.0",
]
